+++ src/zope/testrunner/tests.py
@@ -215,6 +215,7 @@ def test_suite():
         'testrunner-repeat.txt',
         'testrunner-knit.txt',
         'testrunner-shuffle.txt',
+        'testrunner-eggsupport.txt',
         setUp=setUp, tearDown=tearDown,
         optionflags=doctest.ELLIPSIS+doctest.NORMALIZE_WHITESPACE,
         checker=checker),